Fix for AdminMessage InApp message  null pointer issue
This is a simple solution to avoid null pointer exceptions when a user
wants to send an Admin initiated InApp message but doesn’t set the
email subject or email template type.

Currently users are not made aware why the  null pointer exceptions are
thrown and figuring out why is compounded by the fact that the
documentation says these parameters are not required for inapp
messages.

Now the default for email subject is “” and the default for email type
is plain. Of course these are overwritten when specified.